requirements.txt
django
djangorestframework
django-cors-headers
mysqlclient
pillow
django-jazzmin
requests
